requirements.txt
Pillow>=5.3.0
requests>=2.21.0
